+++ app/scripts/migrations/index.js
@@ -37,4 +37,6 @@ module.exports = [
   require('./024'),
   require('./025'),
   require('./026'),
+  require('./027'),
+  require('./028'),
 ]